integer fix because I am God

+++ VAST/core/submodels/output_submodels/vlm_post_processing/horseshoe_circulations.py
@@ -69,7 +69,7 @@ class HorseshoeCirculations(Model):
         cols = np.concatenate(cols)
 
         mtx_val = csc_matrix((data, (rows, cols)),
-                             shape=(system_size, system_size)).toarray()
+                             shape=(system_size, system_size)).astype(int).toarray()
 
         mtx = self.create_input('mtx', val=mtx_val)
